+++ src/test/test.c
@@ -70,10 +70,8 @@ int main(int argc, char ** argv)
 
     struct hash_inputs_statistics hash_inputs_statistics;
     hash_inputs_get_statistics(hash_inputs, &hash_inputs_statistics);
-    printf("[instats] n_growths = %lu\n",
-            (unsigned long)hash_inputs_statistics.n_growths);
-    printf("[instats] capacity = %lu\n",
-            (unsigned long)hash_inputs_statistics.capacity);
+    printf("[instats] n_growths = %zu\n", hash_inputs_statistics.n_growths);
+    printf("[instats] capacity = %zu\n", hash_inputs_statistics.capacity);
 
     f = fopen("keys", "w");
     hash_inputs_apply(hash_inputs, dump_to_file, NULL);
@@ -101,7 +99,8 @@ int main(int argc, char ** argv)
         printf("result2 is null\n");
     }
 
-    const struct hash_lookup_result * result3 = hash_lookup(hash, keep_s, strlen(keep_s));
+    const struct hash_lookup_result * result3 =
+        hash_lookup(hash, keep_s, strlen(keep_s));
     if (result3) {
         printf("found result3\n");
     } else {
